backend/requirements-test.txt
tox==3.12.1
black==19.3b0
pytest==4.6.3
pytest-django==3.5.0
pytest-cov==2.7.1
flake8==3.7.7
parameterized==0.7.0
Sphinx==2.0.1
sphinxcontrib-websupport
sphinx_rtd_theme
freezegun==0.3.12
